backend/requirements.txt
flask
SQLAlchemy~=1.3.20
oso
sqlalchemy-oso-preview
